refactor componente Map
